handled the update and delete opertions of  contact api

+++ controllers/contactController.js
@@ -39,6 +39,10 @@ const updateContact= async (req,res)=>{
         res.status(404)
         throw new Error("Contact not found")
     }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("user dont have permission to updae other user contacts")
+    }
     const updatedContact = await Contact.findByIdAndUpdate(req.params.id,req.body,{new:true})
     res.status(200).json({updatedContact})
 }
@@ -47,6 +51,10 @@ const deleteContact = async(req,res)=>{
     if(!contact){
         res.status(404)
         throw new Error("Contact not found")
+    }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("user dont have permission to delete other user contacts")
     }
     await Contact.findByIdAndDelete(req.params.id)
     res.status(200).json({message:`deleted contact for ${req.params.id}`})
